[Refactor] AppPreviewsView 하위 뷰로 분리
- 앱 프리뷰 이미지 영역 하위 뷰로 분리
